2002-08-18  Bradford Hovinen  <hovinen@cis.udel.edu>
	* examples/blackbox/ex-mat0.C (main): Use Modular<uint32>

	* examples/blackbox/example.C: Use Modular<uint32>

	* examples/fields/ex-fields0.C (main): Use Modular<uint32>

	* examples/fields/ex-fields.C (main): Use Modular<uint32>

	* examples/blackbox/load-symmetrize.C: Use Modular<uint32>

	* examples/blackbox/load-det.C: Use Modular<uint16>

	* examples/dot-product.C: Use Modular<uint32>

	* linbox/field/modular.inl (dotSpecializedDD, dotSpecializedDSP):
	Update signatures of specialization

	* tests/test-trace.C (main): Use Modular<uint32>

	* tests/test-sum.C (main):

	* tests/test-submatrix.C (main):

	* tests/test-sparse.C (main):

	* tests/test-solve.C (main):

	* tests/test-scalar-matrix.C (main):

	* tests/test-minpoly.C (main):

	* tests/test-inverse.C (main):

	* tests/test-diagonal.C (main):

	* tests/test-det.C (main):

	* tests/test-rank.C (main):

	* tests/test-dense.C (main):

	* tests/test-randiter-nonzero.C (main):

	* tests/test-moore-penrose.C (main):

	* tests/test-vector-domain.C: Use Modular<uint*>

	* tests/test-modular.C:

	* linbox/field/modular.h: Change Modular, FieldAXPY, and
	DotProductDomain specializations to use new Linbox uint* types

	* linbox/integer.h: Add typedefs for int*

	* acconfig.h: Add #undef __LINBOX_INT*

	* configure.in: Define __LINBOX_INT* based on results of
	AC_SIZEOF

+++ trunk/linbox/examples/blackbox/load-symmetrize.C
@@ -28,7 +28,7 @@
 using namespace LinBox;
 using namespace std;
 
-typedef Modular<long> Field;
+typedef Modular<uint32> Field;
 
 typedef vector <Field::Element> Vector;
 typedef vector <Field::Element> Polynomial;
